+++ core/board/NstBoardTengenRambo1.cpp
@@ -166,11 +166,11 @@ namespace Nes
 					{
 						const byte data[3] =
 						{
-							(irq.unit.enabled   ? 0x1U : 0x0U) |
+							static_cast<byte>((irq.unit.enabled   ? 0x1U : 0x0U) |
 							(irq.m2.Connected() ? 0x2U : 0x0U) |
-							(irq.unit.reload    ? 0x4U : 0x0U),
-							irq.unit.latch,
-							irq.unit.count & 0xFF
+							(irq.unit.reload    ? 0x4U : 0x0U)),
+							static_cast<byte>(irq.unit.latch),
+							static_cast<byte>(irq.unit.count & 0xFF)
 						};
 
 						state.Begin( AsciiId<'I','R','Q'>::V ).Write( data ).End();
